requirements.txt
langchain==0.3.21
langchain_community==0.3.20
langchain_google_genai==2.1.2
pandas==2.2.3
python-dotenv==1.1.0
streamlit==1.44.0
Werkzeug==3.1.3
faiss-cpu==1.10.0
openpyxl==3.1.5
pypdf==5.4.0

